Move servers list tests to their own file
